+++ development/Sandbox/Program.cs
@@ -6,10 +6,10 @@ Console.Clear();
 //await Sandbox.Providers.OpenAI.RunAsync();
 //await Sandbox.Providers.GitHub.RunAsync();
 //await Sandbox.Providers.Cohere.RunAsync();
-//await AzureOpenAI.RunAsync();
+await AzureOpenAI.RunAsync();
 //await Sandbox.Providers.Mistral.RunAsync();
 //await Sandbox.Providers.Google.RunAsync();
 //await Sandbox.Providers.XAI.RunAsync();
-await Sandbox.Providers.OpenRouter.RunAsync();
+//await Sandbox.Providers.OpenRouter.RunAsync();
 
 Console.WriteLine("Done");