+++ development/Sandbox/Providers/Google.cs
@@ -19,43 +19,17 @@ public static class Google
         //Create your AgentFactory (using a connection object for more options)
         GoogleAgentFactory agentFactory = new GoogleAgentFactory(new GoogleConnection
         {
-            ApiKey = "<apiKey>",
+            ApiKey = secrets.GoogleGeminiApiKey,
         });
 
-        //Create your Agent
-        GoogleAgent agent = agentFactory.CreateAgent(new GoogleAgentOptions
+        AIAgent agent = agentFactory.CreateAgent(new GoogleAgentOptions
         {
-            //Mandatory
-            Model = "gemini-2.5-flash", //Model to use
-
-            //Optional (Common)
-            Name = "MyAgent", //Agent Name
-            MaxOutputTokens = 2000, //Max allow token
-            Temperature = 0, //The Temperature of the LLM Call (1 = Normal; 0 = Less creativity)
-            ThinkingBudget = 5000, //Set Thinking Budget
-            Instructions = "You are a nice AI", //The System Prompt for the Agent to Follow
-            Tools = [], //Add your tools for Tool Calling here
-            ToolCallingMiddleware = async (callingAgent, context, next, token) => //Tool Calling Middleware to Inspect, change, and cancel tool-calling
-            {
-                AIFunctionArguments arguments = context.Arguments; //Details on the tool-call that is about to happen
-                return await next(context, token);
-            },
-            OpenTelemetryMiddleware = new OpenTelemetryMiddleware(source: "MyOpenTelemetrySource", telemetryAgent => telemetryAgent.EnableSensitiveData = true), //Configure OpenTelemetry Middleware
-
-            //Optional (Rarely used)
-            Id = "1234", //Set the ID of Agent (else a random GUID is assigned as ID)
-            Description = "My Description", //Description of the Agent (not used by the LLM)
-            LoggingMiddleware = new LoggingMiddleware( /* Configure custom logging */),
-            Services = null, //Setup Tool Calling Service Injection (See https://youtu.be/EGs-Myf5MB4 for more details)
-            LoggerFactory = null, //Setup logger Factory (Alternative to Middleware)
-            AdditionalChatClientAgentOptions = options =>
-            {
-                //Option to set even more options if not covered by AgentFrameworkToolkit
-            },
-            RawToolCallDetails = Console.WriteLine, //Raw Tool calling Middleware (if you just wish to log what tools are being called. ToolCallingMiddleware is a more advanced version of this)
+            Model = "gemini-3-pro-preview"
         });
+        await agent.RunAsync<string>("");
 
-        AgentRunResponse response = await agent.RunAsync("Hello World");
+
+        AgentRunResponse response = await agent.RunAsync("Why is the Sky Blue");
         Console.WriteLine(response);
     }
 }