+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog - Agent Framework Toolkit
 
+## Unreleased
+- Fixed that the AIAgentExtensions was in a wrongly names class called AIToolsExtensions
+- `.RunAsync<T>(...)` in AIAgentExtensions now used provided serializationOptions (if given); only creating it's own if null.
+- `.RunAsync<T>(...)` now support List/Array as T (using similar object wrapping technique as Microsoft.Extensions.AI)
+- AzureOpenAI: Added Auto-correction of the Endpoint if you by mistake give the entire foundry url 'https://[name].services.ai.azure.com/api/projects/[project]' by on the fly correcting it to 'https://[name].services.ai.azure.com'. This is on by default, but can be turned off by setting property `AutoCorrectFoundryEndpoint` to false
+---
+
 ## Version 1.0.0-preview.260108.1
-- Bumped Agent Framework Nuget Pacakges to '1.0.0-preview.260108.1' and fixed breaking changes (nothing breaking for the Toolkit itself)
+- Bumped Agent Framework Nuget Packages to '1.0.0-preview.260108.1' and fixed breaking changes (nothing breaking for the Toolkit itself)
 
 ---